scripts/requirements.txt
nba_api
pandas
sqlalchemy
psycopg2
python-dotenv